Move multiplexed signals configuration and lookup to peripheral specific header/source file pairs
Resolves #245 (Move multiplexed signals configuration and lookup to
peripheral specific header/source file pairs).
