+++ source/CMakeLists.txt
@@ -31,7 +31,13 @@ set(
     "picolibrary/microchip/megaavr/i2c.cc"
     "picolibrary/microchip/megaavr/multiplexed_signals.cc"
     "picolibrary/microchip/megaavr/multiplexed_signals/atmega2560.cc"
+    "picolibrary/microchip/megaavr/multiplexed_signals/atmega2560/spi.cc"
+    "picolibrary/microchip/megaavr/multiplexed_signals/atmega2560/twi.cc"
+    "picolibrary/microchip/megaavr/multiplexed_signals/atmega2560/usart.cc"
     "picolibrary/microchip/megaavr/multiplexed_signals/atmega328p.cc"
+    "picolibrary/microchip/megaavr/multiplexed_signals/atmega328p/spi.cc"
+    "picolibrary/microchip/megaavr/multiplexed_signals/atmega328p/twi.cc"
+    "picolibrary/microchip/megaavr/multiplexed_signals/atmega328p/usart.cc"
     "picolibrary/microchip/megaavr/peripheral.cc"
     "picolibrary/microchip/megaavr/peripheral/atmega2560.cc"
     "picolibrary/microchip/megaavr/peripheral/atmega328p.cc"
